incorporate flowchart creation engine in genpipes
The flowchart creation engine is incorporated into the core pipeline script such that the flowchart can be automatically created when the pipeline is built without any more input from the user. The code right now works as a minimal working example for the DnaSeq pipeline. More pipelines can be similarly incorporated if the hierarchy between steps is available. Further steps would include improving the implementation, adding proper documentation and a testing framework for continuous integration of new steps into the hierarchy file and thus the flowchart creation engine.